commited many changes:
Done:

login - 15-aug-2023
singup - 15-aug-2023

user verification using email - 16-aug-2023

- sending verification link on email
- verifiying link
forget password
- sending otp on email
- verifiying otp
- user authentication complete (incuding mailing services)

further panning about the databse - 17 aug

- role management - research
- role schema - created
- product management - research

product management api's (crud) - 18 aug

- understand postman
- role based crud of products

redux toolkit implementation (signin, signup, auth slices only) - 19 aug
login, signup, Header - ui

replaced jwt with josh web token - 20 aug
made new page cluster (user pages ui) - 20 aug
added orders crud in backend - 20 aug
changed the role object in roles and permissions table.

cart crud in backend + ui - 21 aug

converting only client components to server + client components - 22 aug

to do:
complete UI + integration from backend  ( goal )
target 1:  wistlist crud in backend

shortcuts learned:
shift + num_8 => volume up
shift + num_2 => volume down
f3, f4 => previous, next track
super + L => sleep
cntrl + L => select line

+++ eshop/app/api/products/[productId]/route.js
@@ -3,13 +3,13 @@ import { NextResponse } from 'next/server';
 import Products from '@/models/Products';
 import Users from '@/models/Users';
 import Roles from '@/models/Roles';
+import { getUserfromJwt, isPermitted } from '@/utils/auth/auth';
 
 
 export const GET = async (request, response) => {
   try {
     await connect();
-    // console.log("request.url: ", request.url);
-    // console.log("response: ", response);
+
     const productId = response.params.productId;
     if (!productId) {
       return NextResponse.json({
@@ -43,32 +43,26 @@ export const GET = async (request, response) => {
 export const DELETE = async (request, response) => {
   try {
     await connect();
-    const productId = response.params.productId;
 
+    const { email } = await getUserfromJwt(request.cookies.get('token').value);
+    const user = await Users.findOne({ email });
+    const role = await Roles.findOne({ name: user.role });
+    const reqPermission = [{ resource: 'products', actions: ['delete'] }];
+    const authorized = isPermitted(reqPermission, role.permissions);
+    if (!authorized) {
+      return NextResponse.json({
+        status: 400,
+        body: { message: 'Not permitted', success: false }
+      });
+    }
+
+    const productId = response.params.productId;
     const deletedProduct = await Products.findOneAndDelete({ _id: productId });
 
     if (!deletedProduct) {
       return NextResponse.json({
         status: 404,
         body: { message: 'Product not found', success: false }
-      });
-    }
-
-    const shopper = await Users.findOne({ _id: deletedProduct.shopperId });
-    if (!shopper) {
-      return NextResponse.json({
-        status: 400,
-        body: { message: 'Shopper not found', success: false }
-      });
-    }
-
-    const role = await Roles.findOne({ _id: shopper.role });
-    const isPermitted = role.permissions.some(permission =>
-      permission.resource === 'products' && permission.actions.includes('delete'));
-    if (!isPermitted) {
-      return NextResponse.json({
-        status: 400,
-        body: { message: 'Not permitted', success: false }
       });
     }
